bot-server/requirements.txt
python-telegram-bot==21.10
aiohttp==3.11.18
